Move rebecca/fakedriver to rebecca/driver/fake
